creating test helper function windowRectangleToCanvasRectangle
